add factory method for parsers
